+++ frontend/lib/constants/route.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart' as material;
 import 'package:paotung_frontend/screens/main/account/presentation/add_transaction.dart';
 import 'package:paotung_frontend/screens/main/profile/newcategory.dart';
-import 'package:paotung_frontend/screens/main/profile/categorysetting.dart';
+import 'package:paotung_frontend/screens/main/profile/category_setting.dart';
 import 'package:paotung_frontend/screens/main/profile/notification_setting.dart';
 import 'package:paotung_frontend/screens/profile/help_center.dart';
 import 'package:paotung_frontend/screens/splashscreen/splash.dart';
